fix: remove thread pool loop
